requirements.txt
django
djangorestframework
django-environ
flake8
qrcode
fpdf


